Register each model class as a JAX pytree.
This allows them to be manipulated and passed through JAX transforms,
such as mapping over models or passing them as arguments to JIT
compiled functions. We also only store small constants and parameters
as attributes of the classes. Other values are recomputed on demand.
The intent is that the outermost use of these models should be JIT
compiled allowing JAX to observe the structure of these computations,
eliminate repeated calculations, and store them as constants if
advantageous.