Fix breakage due to eval_shape now trying to weakref callables.
It seems that JAX v0.4.24 changed the implementation of eval_shape to
internally use the JIT. This tries to weak reference the provided
callable for a cache. The operator.attrgetter objects we previously
used are not weak-referenceable types. To fix, replace them with
lambdas. One use also adds a functools.partial wrapped to silence a
lint warning about not binding the name variable from the outer loop.