Update copyright notices to use standard format.
This package was first published and made available in 2023. We take
this as our year of first publication and use it across all current
copyright notices except for those files added in later years. Later
modifications may justify incrementing this year, but this should be
good enough for our purposes.